LPS-76837 Rename test module
