--- modules/apps/foundation/text-localizer/text-localizer-address-test/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Text Localizer Address Test
-Bundle-SymbolicName: com.liferay.text.localizer.address.test
-Bundle-Version: 1.0.0
--includeresource: test-classes/integration
