+++ modules/apps/foundation/text-localizer/text-localizer-taglib-test/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Text Localizer Taglib Test
+Bundle-SymbolicName: com.liferay.text.localizer.taglib.test
+Bundle-Version: 1.0.0
+-includeresource: test-classes/integration
