+++ modules/apps/foundation/text-localizer/text-localizer-taglib-test/build.gradle
