+++ modules/apps/foundation/text-localizer/text-localizer-taglib-test/src/testIntegration/java/com/liferay/text/localizer/taglib/util/test/AddressTextLocalizerUtilTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.text.localizer.address.util.test;
+package com.liferay.text.localizer.taglib.util.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.portal.kernel.model.Address;
